Move tests close to their source files
