+++ src/bin.ts
@@ -2,7 +2,6 @@
 
 import * as yargs from 'yargs';
 import chalk from 'chalk';
-import { compareExports, areChangesBreaking } from './commands/compare';
 import { getImportsInfo, getGroupedImports } from './compiler/imports';
 import { getListImportsCliArgs, CliError } from './utils/cli';
 import { resolvePackage, resolveTargetPackages } from './utils/npm';
@@ -10,10 +9,11 @@ import { getExportInfo } from './compiler/exports';
 import { exit } from 'process';
 import { access } from 'fs/promises';
 import { constants } from 'fs';
-import { isCompatible } from './commands/is-compatible';
-import { printComparison } from './print/comparison';
 import { printImports } from './print/imports';
 import { printExports } from './print/exports';
+import { areChangesBreaking, compareExports } from './commands/compare/compare';
+import { printComparison } from './print/comparison';
+import { isCompatible } from './commands/is-compatible/is-compatible';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions
 yargs
