+++ src/print/comparison.ts
@@ -1,5 +1,4 @@
 import chalk from 'chalk';
-import { areChangesBreaking } from '../commands/compare';
 import { Comparison, Exports } from '../types';
 import { debug } from '../utils/log';
 import { printChanges } from './changes';
@@ -7,6 +6,7 @@ import { printRemovals } from './removals';
 import { printHeading, printSpacing } from './utils';
 import { printVerdict } from './veredict';
 import Table from 'tty-table';
+import { areChangesBreaking } from '../commands/compare/compare';
 
 export function printComparison({ changes, additions, removals, prevProgram, currentProgram }: Comparison) {
   debug('Printing results...');
